refactor : 테스트 구조 개선
