+++ src/test/java/gift/member/MemberTest.java
@@ -1,0 +1,5 @@
+package gift.member;
+
+public class MemberTest {
+
+}
